Common code for KeyBindings
