+++ fabric/src/main/java/club/iananderson/seasonhud/client/SeasonHUDClientFabric.java
@@ -2,6 +2,7 @@ package club.iananderson.seasonhud.client;
 
 import club.iananderson.seasonhud.Common;
 import club.iananderson.seasonhud.config.Config;
+import club.iananderson.seasonhud.event.ClientEvents;
 import fuzs.forgeconfigapiport.api.config.v2.ForgeConfigRegistry;
 import net.fabricmc.api.ClientModInitializer;
 import net.minecraftforge.fml.config.ModConfig;
@@ -11,7 +12,7 @@ public class SeasonHUDClientFabric implements ClientModInitializer {
   @Override
   public void onInitializeClient() {
     ForgeConfigRegistry.INSTANCE.register(Common.MOD_ID, ModConfig.Type.CLIENT, Config.GENERAL_SPEC,
-        "SeasonHUD-client.toml");
-    KeyBindings.register();
+                                          "SeasonHUD-client.toml");
+    ClientEvents.register();
   }
 }
